+++ CHANGELOG.md
@@ -13,8 +13,9 @@
 -   Feature: expose `@Debounce()` decorator for throttling dispatch actions
 -   Feature: add extension API for NGXS Data plugin
 -   Feature: expose storage extension as plugin
--   Feature: add fire init option to `@Persistence()` (storage extension)
--   Feature: improved `@Persistence()` decorator for stability
+-   Feature: add fire init option to `@Persistence()`
+-   Feature: add rehydrate option to `@Persistence()`
+-   Fix: improved `@Persistence()` decorator for stability
 -   Fix: use deep equals when state update in storage from another browser tab
 -   Fix: now `@Computed()` fields are recalculation when store has changed
 -   Fix: correct recalculation of state during inheritance computed fields